Update Bank.sol
checks if the sender of a message (msg.sender) already has an account (customerExist[msg.sender]) and throws an error if they do ("You already have an account"). If the sender does not have an account, it sets the accountCreated.UserAddr to the msg.sender and sets customerExist[msg.sender] to true, indicating that the sender now has an account.